[tested] move plugin interface to core
