+++ pkg/plugins/filetypefilter.go
@@ -8,6 +8,7 @@ import (
 	"path/filepath"
 	"slices"
 
+	"github.com/gardener/docforge/pkg/core"
 	"github.com/gardener/docforge/pkg/core/manifest"
 	"github.com/gardener/docforge/pkg/core/registry"
 )
@@ -50,7 +51,7 @@ func (p *FileTypeFilterPlugin) Process(*manifest.Node) error {
 }
 
 // ProcessNew processes a node using the new channel-based method
-func (p *FileTypeFilterPlugin) ProcessNew(*manifest.Node) []chan Status {
+func (p *FileTypeFilterPlugin) ProcessNew(*manifest.Node) []chan core.Status {
 	return nil // Not used
 }
